+++ src/Nethermind/Nethermind.Merge.Plugin/BlockProduction/IBlockProductionContext.cs
@@ -4,7 +4,7 @@
 using Nethermind.Core;
 using Nethermind.Int256;
 
-namespace Nethermind.Consensus;
+namespace Nethermind.Merge.Plugin.BlockProduction;
 
 public interface IBlockProductionContext
 {
