+++ src/Nethermind/Nethermind.Merge.Plugin/BlockProduction/IPayloadPreparationService.cs
@@ -3,7 +3,6 @@
 
 using System;
 using System.Threading.Tasks;
-using Nethermind.Consensus;
 using Nethermind.Consensus.Producers;
 using Nethermind.Core;
