+++ src/Nethermind/Nethermind.Merge.Plugin/BlockProduction/NoBlockImprovementContext.cs
@@ -3,7 +3,6 @@
 
 using System;
 using System.Threading.Tasks;
-using Nethermind.Consensus;
 using Nethermind.Core;
 using Nethermind.Int256;
